requirements.txt
aiogram==3.4.1
openai>=1.12.0
faiss-cpu>=1.9.0
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.26.3
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
rank_bm25>=0.2.2
sentence-transformers>=2.2.0
